ShashChess 33.1 Fix learning. Now, the experience file has a .exp extension and not .bin as before. So, simply, rename the file. Stockfish patch Update default net to nn-5af11540bbfe.nnue

+++ src/All/thread.h
@@ -78,7 +78,7 @@ public:
   ContinuationHistory continuationHistory[2][2];
   bool fullSearch;//full threads patch
   //begin from Shashin
-  int shashinWinProbabilityRange=0;
+  uint8_t shashinWinProbabilityRange=0;
   Key shashinPosKey;
   Depth shashinDepth=0;
   //end from Shashin
